Split test_fitbit_routes: Move tests for multiple notification webhook calls to a separate file.
